Numerous fixes to PG type name handling (#1945)
* Return the standard, user-facing type names rather than the legacy
  names PostgreSQL uses internally
  (https://www.postgresql.org/docs/current/static/datatype.html#DATATYPE-TABLE).
  For example, we need to return `integer` rather than `int4`, and for
  arrays we need to return `text[]` rather than the internal `_text`.
  This should happen everywhere.
* Include facet information in `NpgsqlDataReader.GetDataTypeName()`,
  returning `numeric(x,y)` instead of just `numeric`.
* Column information returned by `NpgsqlDataReader.GetColumnSchema()`
  sometimes did not correctly populate the facets
  (size/precision/scale).

Fixes #1919

+++ src/Npgsql/TypeHandlers/NumericHandlers/Int64Handler.cs
@@ -36,7 +36,7 @@ namespace Npgsql.TypeHandlers.NumericHandlers
     /// <remarks>
     /// http://www.postgresql.org/docs/current/static/datatype-numeric.html
     /// </remarks>
-    [TypeMapping("int8", NpgsqlDbType.Bigint, DbType.Int64, typeof(long))]
+    [TypeMapping("bigint", NpgsqlDbType.Bigint, DbType.Int64, typeof(long))]
     class Int64Handler : NpgsqlSimpleTypeHandler<long>,
         INpgsqlSimpleTypeHandler<byte>, INpgsqlSimpleTypeHandler<short>, INpgsqlSimpleTypeHandler<int>,
         INpgsqlSimpleTypeHandler<float>, INpgsqlSimpleTypeHandler<double>, INpgsqlSimpleTypeHandler<decimal>,
